+++ sink/sinker.go
@@ -16,6 +16,7 @@ import (
 	"github.com/streamingfast/bstream"
 	"github.com/streamingfast/derr"
 	"github.com/streamingfast/dgrpc"
+	"github.com/streamingfast/dmetrics"
 	"github.com/streamingfast/shutter"
 	"github.com/streamingfast/substreams/client"
 	pbsubstreamsrpc "github.com/streamingfast/substreams/pb/sf/substreams/rpc/v2"
@@ -169,9 +170,9 @@ func (s *Sinker) ApiToken() string {
 func (s *Sinker) PrintStats() {
 	fmt.Fprintln(os.Stderr, "")
 	fmt.Fprintln(os.Stderr, "📊 Usage Report")
-	fmt.Fprintf(os.Stderr, " • Egress Bytes (uncompressed): %s\n", humanize.IBytes(uint64(uint64(DataMessageSizeBytes.Get()))))
-	fmt.Fprintf(os.Stderr, " • Processed Blocks: %s blocks\n", humanize.Comma(int64(ProgressMessageTotalProcessedBlocks.Get())))
-	fmt.Fprintf(os.Stderr, " • Processed Bytes: %s\n", humanize.IBytes(uint64(ProgressMessageProcessedBytes.Get())))
+	fmt.Fprintf(os.Stderr, " • Egress Bytes (uncompressed): %s\n", humanize.IBytes(uint64(uint64(ServerEgressBytes.Get()))))
+	fmt.Fprintf(os.Stderr, " • Processed Blocks: %s blocks\n", humanize.Comma(int64(ProcessedBlocks.Get())))
+	fmt.Fprintf(os.Stderr, " • Processed Bytes: %s\n", humanize.IBytes(uint64(ProcessedBytes.Get())))
 }
 
 func (s *Sinker) Run(ctx context.Context, cursor *Cursor, handler SinkerHandler) {
@@ -182,6 +183,10 @@ func (s *Sinker) Run(ctx context.Context, cursor *Cursor, handler SinkerHandler)
 		cancel()
 	})
 	s.stats.OnTerminated(func(err error) { s.Shutdown(err) })
+	if s.SinkerConfig.PrometheusAddr != "" {
+		Metrics.Register()
+		go dmetrics.Serve(s.SinkerConfig.PrometheusAddr)
+	}
 
 	logEach := 15 * time.Second
 	if s.Logger.Core().Enabled(zap.DebugLevel) {
@@ -463,8 +468,8 @@ func (s *Sinker) doRequest(
 			ProgressMessageCount.Inc()
 			// The returned value from the server gives an overview of the current progress and not the delta
 			// since the last message. Since the server is the source of truth, we just set the value directly.
-			ProgressMessageTotalProcessedBlocks.SetUint64(r.Progress.ProcessedBlocks)
-			ProgressMessageProcessedBytes.SetUint64(r.Progress.ProcessedBytes.TotalBytesRead)
+			ProcessedBlocks.SetUint64(r.Progress.ProcessedBlocks)
+			ProcessedBytes.SetUint64(r.Progress.ProcessedBytes.TotalBytesRead)
 
 			if s.Tracer.Enabled() {
 				s.Logger.Debug("received response Progress", zap.Reflect("progress", r))
@@ -495,7 +500,7 @@ func (s *Sinker) doRequest(
 			HeadBlockNumber.SetUint64(block.Num())
 			HeadBlockTimeDrift.SetBlockTime(r.BlockScopedData.Clock.Timestamp.AsTime())
 			DataMessageCount.Inc()
-			DataMessageSizeBytes.AddInt(proto.Size(r.BlockScopedData))
+			ServerEgressBytes.AddInt(proto.Size(r.BlockScopedData))
 			BackprocessingCompletion.SetUint64(1)
 
 			cursor, err := NewCursor(r.BlockScopedData.Cursor)